requirements.txt
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
allure-pytest==2.13.3
allure-python-commons==2.13.3
altgraph==0.17.4
annotated-doc==0.0.3
annotated-types==0.7.0
ansible-core==2.17.13
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backports.asyncio.runner==1.2.0
bcc==0.18.0
bcrypt==4.0.1
blinker==1.4
Brlapi==0.8.3
catkin-pkg==1.1.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.4.3
chrome-gnome-shell==0.0.0
click==8.3.0
click-completion==0.5.2
click-help-colors==0.9.4
click-repl==0.3.0
click-spinner==0.1.10
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==23.10.4
contourpy==1.3.2
cryptography==46.0.1
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debtcollector==3.0.0
deepmerge==1.1.1
defer==1.0.6
dictdiffer==0.9.0
dill==0.4.0
directory_tree==1.0.0
distlib==0.4.0
distro==1.7.0
dnspython==2.7.0
docutils==0.22.2
ecdsa==0.19.1
email-validator==2.0.0.post2
etcd3gw==2.4.2
exceptiongroup==1.3.0
Faker==37.8.0
fastapi==0.120.2
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.19.1
filterpy==1.4.5
fonttools==4.59.2
frozenlist==1.7.0
futurist==3.2.0
gitdb==4.0.12
GitPython==3.1.43
graphlib_backport==1.1.0
graphviz==0.21
greenlet==3.2.4
gyp==0.1
h11==0.16.0
hidpidaemon==18.4.6
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.13
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
iniconfig==2.1.0
invoke==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.6.4
munch==4.0.0
nested-lookup==0.2.25
netaddr==0.8.0
netifaces==0.11.0
nodeenv==1.9.1
numpy==1.24.3
oauthlib==3.2.0
oks-utils==0.2.19.dev2
packaging==25.0
paho-mqtt==1.6.1
pandas==2.0.3
paramiko==4.0.0
passlib==1.7.4
pbr==7.0.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
polling2==0.5.0
pop-transition==1.1.2
pre_commit==4.3.0
pretty-traceback==2024.1021
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.9
pydantic_core==2.33.2
pydbus==0.6.0
Pygments==2.19.2
PyGObject==3.42.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.2
PyNaCl==1.5.0
pyOpenSSL==25.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-benchmark==4.0.0
pytest-env==1.1.5
pytest-repeat==0.9.4
pytest-reportportal==5.5.2
python-apt==2.4.0+ubuntu4
python-benedict==0.34.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-fsutil==0.15.0
python-gnupg==0.4.8
python-hosts==1.1.2
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-on-whales==0.73.0
python-slugify==8.0.4
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
rapyuta-io-cli==9.12.2
rapyuta_io==3.0.0
referencing==0.36.2
repolib==2.2.1
repoman==1.4.0
reportportal-client==5.6.5
requests==2.32.5
resolvelib==1.0.1
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.2
roslibpy==1.7.0
rospkg==1.6.0
rpds-py==0.27.0
rsa==4.9.1
scipy==1.15.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semver==3.0.4
sentry-sdk==2.43.0
service-identity==24.2.0
sessioninstaller==0.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
ssh-import-id==5.11
starlette==0.49.1
StrEnum==0.4.15
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
terminator==2.1.1
text-unidecode==1.3
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.34.0
wadllib==1.3.6
waiting==1.5.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.3
xdg==5
xkit==0.0.0
yarl==1.20.1
yaspin==2.5.0
zipp==1.0.0
zope.interface==8.0.1
